Resume_Screener/backend/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
pdfplumber
python-docx
PyPDF2
